backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
distro==1.9.0
Django==5.1.2
django-cors-headers==4.5.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
elevenlabs==1.10.0
etelemetry==0.3.1
filelock==3.16.1
fitz==0.0.1.dev2
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
isodate==0.6.1
jiter==0.6.1
looseversion==1.3.0
lxml==5.3.0
networkx==3.4.1
nibabel==5.3.1
nipype==1.8.6
numpy==2.1.2
openai==1.52.0
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
prov==2.0.1
PyAudio==0.2.14
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
scipy==1.14.1
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
traits==6.3.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websockets==13.1
